+++ src/main/java/com/example/miniredditbackend/post/PostSer.java
@@ -8,13 +8,13 @@ public interface PostSer {
 
     Posts createPost(Posts posts, String token);
     List<PostDTO> getAllPosts();
-    void voteUp(String token, int id);
-    void voteDown(String token, int id);
-    int createComment(String commentAuthor, String comment, String date, int id, String token);
-    boolean deletePost(String token, int id);
-    boolean editPost(String token, int id, String comment, String title);
-    boolean deleteComment(String token, int id);
-    boolean editComment(String token, int idcomment, String comment);
-    PostDTO getPost(int x);
-    List<commentDTO> getComments(int x);
+    void voteUp(String token, long id);
+    void voteDown(String token, long id);
+    int createComment(String commentAuthor, String comment, String date, long id, String token);
+    boolean deletePost(String token, long id);
+    boolean editPost(String token, long id, String comment, String title);
+    boolean deleteComment(String token, long id);
+    boolean editComment(String token, long idcomment, String comment);
+    PostDTO getPost(long x);
+    List<commentDTO> getComments(long x);
 }
